+++ packages/tswebextension/CHANGELOG.md
@@ -34,6 +34,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [AdguardBrowserExtension#3329]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3329
 [AdguardBrowserExtension#3414]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3414
 
+## [4.0.6] - 2026-03-26
+
+### Changed
+
+- Updated [@adguard/agtree] to `v4.0.4`.
+- Updated [@adguard/tsurlfilter] to `v4.0.5`.
+
+[4.0.6]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v4.0.6
+
 ## [4.0.5] - 2026-03-24
 
 ### Changed